pyproject.toml
[project]
name = "malpanez-security"
version = "1.0.0"
description = "TOP 0.1% Enterprise Ansible Security Collection"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Miguel Alpañez", email = "miguel@alpanez.com"}
]
keywords = ["ansible", "security", "compliance", "automation", "hardening"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "ansible-core>=2.19,<2.21",
]

[project.optional-dependencies]
dev = [
    "ansible-lint>=24.5.0",
    "yamllint>=1.35.1",
    "molecule>=25.1.0",
    "molecule-plugins[docker]>=25.8.12",
    "pytest>=8.0.0",
    "pytest-testinfra>=10.0.0",
    "hypothesis>=6.100.0",
    "pre-commit>=4.0.0",
    "ruff>=0.1.0",
]

test = [
    "pytest>=8.0.0",
    "pytest-testinfra>=10.0.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "hypothesis>=6.100.0",
]

lint = [
    "ansible-lint>=24.5.0",
    "yamllint>=1.35.1",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

all = [
    "malpanez-security[dev,test,lint,docs]",
]

[project.urls]
Homepage = "https://github.com/malpanez/security"
Documentation = "https://github.com/malpanez/security/blob/main/README.md"
Repository = "https://github.com/malpanez/security"
Issues = "https://github.com/malpanez/security/issues"
Changelog = "https://github.com/malpanez/security/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["roles", "playbooks"]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    ".venv",
    ".ansible",
    "ansible_collections",
    ".devcontainer",
    ".github",
    "**/.molecule",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # bandit (security)
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "S603",  # subprocess without shell=True (false positive)
    "S607",  # starting process with partial path (false positive)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "I001", "F401"]  # assert, imports, unused imports allowed in tests
"**/tests/**" = ["S101", "I001", "F401"]  # assert, imports, unused imports allowed in molecule tests
"**/test_*.py" = ["S101", "I001", "F401"]  # assert, imports, unused imports allowed in test files

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=roles",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
branch = true
parallel = true
source = ["roles"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
